refactor(XMLHttpRequest): Move implementation to `xhr` subdirectory
