+++ lib/jsdom/browser/Window.js
@@ -13,7 +13,7 @@ const postMessage = require("../living/post-message");
 const DOMException = require("domexception/webidl2js-wrapper");
 const { btoa, atob } = require("abab");
 const idlUtils = require("../living/generated/utils");
-const createXMLHttpRequest = require("../living/xmlhttprequest");
+const createXMLHttpRequest = require("../living/xhr/XMLHttpRequest-impl");
 const WebSocketImpl = require("../living/websockets/WebSocket-impl").implementation;
 const BarProp = require("../living/generated/BarProp");
 const Document = require("../living/generated/Document");
