+++ lib/jsdom/living/xhr/XMLHttpRequest-impl.js
@@ -10,24 +10,24 @@ const conversions = require("webidl-conversions");
 
 const xhrUtils = require("./xhr-utils");
 const DOMException = require("domexception/webidl2js-wrapper");
-const xhrSymbols = require("./xmlhttprequest-symbols");
-const { addConstants } = require("../utils");
-const { documentBaseURLSerialized } = require("./helpers/document-base-url");
-const { asciiCaseInsensitiveMatch } = require("./helpers/strings");
-const idlUtils = require("./generated/utils");
-const Document = require("./generated/Document");
-const Blob = require("./generated/Blob");
-const FormData = require("./generated/FormData");
-const XMLHttpRequestEventTarget = require("./generated/XMLHttpRequestEventTarget");
-const XMLHttpRequestUpload = require("./generated/XMLHttpRequestUpload");
-const ProgressEvent = require("./generated/ProgressEvent");
-const { isArrayBuffer } = require("./generated/utils");
-const { parseIntoDocument } = require("../browser/parser");
-const { fragmentSerialization } = require("./domparsing/serialization");
-const { setupForSimpleEventAccessors } = require("./helpers/create-event-accessor");
-const { parseJSONFromBytes } = require("./helpers/json");
-const { fireAnEvent } = require("./helpers/events");
-const { copyToArrayBufferInNewRealm } = require("./helpers/binary-data");
+const xhrSymbols = require("./XMLHttpRequest-symbols");
+const { addConstants } = require("../../utils");
+const { documentBaseURLSerialized } = require("../helpers/document-base-url");
+const { asciiCaseInsensitiveMatch } = require("../helpers/strings");
+const idlUtils = require("../generated/utils");
+const Document = require("../generated/Document");
+const Blob = require("../generated/Blob");
+const FormData = require("../generated/FormData");
+const XMLHttpRequestEventTarget = require("../generated/XMLHttpRequestEventTarget");
+const XMLHttpRequestUpload = require("../generated/XMLHttpRequestUpload");
+const ProgressEvent = require("../generated/ProgressEvent");
+const { isArrayBuffer } = require("../generated/utils");
+const { parseIntoDocument } = require("../../browser/parser");
+const { fragmentSerialization } = require("../domparsing/serialization");
+const { setupForSimpleEventAccessors } = require("../helpers/create-event-accessor");
+const { parseJSONFromBytes } = require("../helpers/json");
+const { fireAnEvent } = require("../helpers/events");
+const { copyToArrayBufferInNewRealm } = require("../helpers/binary-data");
 
 const syncWorkerFile = require.resolve ? require.resolve("./xhr-sync-worker.js") : null;
