+++ lib/jsdom/living/xhr/XMLHttpRequest-symbols.js
