+++ lib/jsdom/living/xhr/xhr-sync-worker.js
@@ -1,9 +1,9 @@
 "use strict";
 /* eslint-disable no-process-exit */
 const util = require("util");
-const { JSDOM } = require("../../..");
+const { JSDOM } = require("../../../..");
 const tough = require("tough-cookie");
-const xhrSymbols = require("./xmlhttprequest-symbols.js");
+const xhrSymbols = require("./XMLHttpRequest-symbols.js");
 
 const dom = new JSDOM();
 const xhr = new dom.window.XMLHttpRequest();
