+++ lib/jsdom/living/xhr/xhr-utils.js
@@ -6,12 +6,12 @@ const { URL } = require("whatwg-url");
 const parseDataURL = require("data-urls");
 const DOMException = require("domexception/webidl2js-wrapper");
 
-const ProgressEvent = require("./generated/ProgressEvent");
-const idlUtils = require("./generated/utils");
+const ProgressEvent = require("../generated/ProgressEvent");
+const idlUtils = require("../generated/utils");
 
-const xhrSymbols = require("./xmlhttprequest-symbols");
-const wrapCookieJarForRequest = require("./helpers/wrap-cookie-jar-for-request");
-const { fireAnEvent } = require("./helpers/events");
+const xhrSymbols = require("./XMLHttpRequest-symbols");
+const wrapCookieJarForRequest = require("../helpers/wrap-cookie-jar-for-request");
+const { fireAnEvent } = require("../helpers/events");
 
 const headerListSeparatorRegexp = /,[ \t]*/;
 const simpleMethods = new Set(["GET", "HEAD", "POST"]);
